% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_category_comp.R
\name{auc_category_comp}
\alias{auc_category_comp}
\title{Compare the AUC of our model with previously published models for predicting ICI response}
\usage{
auc_category_comp(object, object2, model_name, dataset_col = NULL, dataset)
}
\arguments{
\item{object}{Output of function ML.Dev.Pred.Category.Sig}

\item{object2}{Output of function cal_auc_previous_sig}

\item{model_name}{Method name in object used to compare}

\item{dataset_col}{If NULL, color values are set to the default colors. Otherwise, you can specify consistent number of color value for datasets}

\item{dataset}{A vector of names for all datasets}
}
\value{
a ggplot2 object
}
\description{
Creates a distribution plot of AUC among different mdoels
}
\examples{
auc_category_comp(res.ici,
            auc.other.pre,
            model_name="svmRadialWeights",
            dataset=names(list_train_vali_Data))


}
