% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_vis_category_all.R
\name{auc_vis_category_all}
\alias{auc_vis_category_all}
\title{Title}
\usage{
auc_vis_category_all(object, dataset, order = NULL, method_col = NULL)
}
\arguments{
\item{method_col}{}
}
\description{
Title
}
