% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.auc.category.model.R
\name{cal.auc.category.model}
\alias{cal.auc.category.model}
\title{Title}
\usage{
cal.auc.category.model(res.by.ML.Dev.Pred.Category.Sig, cohort.for.cal)
}
\arguments{
\item{cohort.for.cal}{}
}
\description{
Title
}
