% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_AUC_ml_res.R
\name{cal_AUC_ml_res}
\alias{cal_AUC_ml_res}
\title{Calculate AUC scores of Machine Learning Models in training data}
\usage{
cal_AUC_ml_res(
  res.by.ML.Dev.Prog.Sig = NULL,
  train_data,
  inputmatrix.list,
  mode = NULL,
  AUC_time = 1,
  single_ml = NULL,
  double_ml1 = NULL,
  double_ml2 = NULL,
  alpha_for_Enet = NULL,
  direction_for_stepcox = NULL,
  auc_cal_method = "KM"
)
}
\arguments{
\item{res.by.ML.Dev.Prog.Sig}{the results of function ML.Dev.Prog.Sig}

\item{train_data}{the training data using in ML.Dev.Prog.Sig}

\item{inputmatrix.list}{A list contain the dataframes (colnames:ID,OS.time,OS,other genes), log2(x+1)， OS.time(day), OS(0/1)}

\item{mode}{Choose MF models: all, single, double}

\item{AUC_time}{c(1,2,3,4,5,6,7,······), for 1 year, 2 years, 3 years......We recommend using the shortest survival time among all queues.}

\item{single_ml}{If the mode is set to "single", you must fill in the following models: c("RSF", "Enet", "StepCox","CoxBoost","plsRcox","superpc","GBM","survivalsvm","Ridge","Lasso").}

\item{double_ml1}{If the mode is set to "double", you need to fill in the modeling methods here: c('RSF', "StepCox","CoxBoost","Lasso").}

\item{double_ml2}{If the mode is set to "double", you need to fill in the modeling methods here: c("RSF", "Enet", "StepCox","CoxBoost","plsRcox","superpc","GBM","survivalsvm","Ridge","Lasso")}

\item{alpha_for_Enet}{0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9}

\item{direction_for_stepcox}{c("both", "backward", "forward")}

\item{auc_cal_method}{KM, NNE}
}
\value{
AUC Calculation Result
}
\description{
Calculate AUC scores of Machine Learning Models in training data
}
\examples{
all.auc.1y <- cal_AUC_ml_res(
  res.by.ML.Dev.Prog.Sig = res,
  train_data = list_train_vali_Data[["TCGA"]],
  inputmatrix.list = list_train_vali_Data, mode = "all",
  AUC_time = 1,
  auc_cal_method = "KM"
)

}
