% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_auc_pre.prog.sig.R
\name{cal_auc_pre.prog.sig}
\alias{cal_auc_pre.prog.sig}
\title{Calculate AUC scores of the previous signarues in all data}
\usage{
cal_auc_pre.prog.sig(
  use_your_own_collected_sig,
  collected_sig_table,
  type.sig,
  list_input_data,
  AUC_time = NULL,
  auc_cal_method = "KM"
)
}
\arguments{
\item{use_your_own_collected_sig}{Whether to use your own collected signatures. T or F.}

\item{collected_sig_table}{If use_your_own_collected_sig set as T, you should provide a data frame containing the information of the signatures. The column names of the data frame are "model"  "PMID"   "Cancer" "Author" "Coef"   "symbol". For example, 'Chen.33591634' '33591634''LGG' 'Chen' '0.7426' 'FGF7'. The 'model' consists of the first name of the first author and the PMID. The 'PMID' is from the paper. The 'Cancer' uses abbreviations in the format of the TCGA. 'Author' is the first name of the first author of the paper. 'Coef' is the coefficient of the variable in the signature. 'symbol' is the variable in the signature. If use_your_own_collected_sig is F, you don't need to provide this data.}

\item{type.sig}{If the use_your_own_collected_sig is F, here we collected some signatures of the LGG, GBM, and Glioma. You can choose c('Glioma', 'LGG', 'GBM'), c('Glioma'),c('GBM'), c('Glioma', 'LGG'), and some other combination for the signatures you want.}

\item{list_input_data}{A list of the cohorts. Column names are 'ID', 'OS.time', 'OS', and the other variables. OS.time (Day). OS(1/0).}

\item{AUC_time}{c(1,2,3,4,5,6,7,······), for 1 year, 2 years, 3 years......We recommend using the shortest survival time among all queues.}

\item{auc_cal_method}{'KM', 'NNE'. The default is 'KM'.}
}
\value{
A list of the AUC results of each previous signature in each cohort you provide.
}
\description{
Calculate AUC scores of the previous signarues in all data
}
