% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_cindex_pre.prog.sig.R
\name{cal_cindex_pre.prog.sig}
\alias{cal_cindex_pre.prog.sig}
\title{Title}
\usage{
cal_cindex_pre.prog.sig(
  use_your_own_collected_sig,
  collected_sig_table,
  type.sig,
  list_input_data
)
}
\arguments{
\item{list_input_data}{}
}
\description{
Title
}
