% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_RS_ml_res.R
\name{cal_RS_ml_res}
\alias{cal_RS_ml_res}
\title{Title}
\usage{
cal_RS_ml_res(
  res.by.ML.Dev.Prog.Sig = NULL,
  train_data,
  inputmatrix.list,
  mode = NULL,
  single_ml = NULL,
  double_ml1 = NULL,
  double_ml2 = NULL,
  alpha_for_Enet = NULL,
  direction_for_stepcox = NULL
)
}
\arguments{
\item{direction_for_stepcox}{}
}
\description{
Title
}
