% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_unicox_ml_res.R
\name{cal_unicox_ml_res}
\alias{cal_unicox_ml_res}
\title{Title}
\usage{
cal_unicox_ml_res(
  res.by.ML.Dev.Prog.Sig = NULL,
  optimal.model,
  type = "categorical"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
