% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuno_heatmap.R
\name{immuno_heatmap}
\alias{immuno_heatmap}
\title{Compare immune cell infiltration between high-risk and low-risk group determined by specific model}
\usage{
immuno_heatmap(object, object2, model_name, col = NULL, dataset)
}
\arguments{
\item{object}{Output of function ML.Dev.Prog.Sig}

\item{object2}{Output of function TME_deconvolution_all}

\item{model_name}{Model name in object used to compare}

\item{col}{If NULL, color values are set to the default colors. Otherwise, you can specify two color values for high and low risk group}

\item{dataset}{Names for specific dataset}
}
\value{
a heatmap object
}
\description{
Creates a heatmap of immune cell infiltration among different methods
}
\examples{
immuno_heatmap(res,
            devo,
            model_name="StepCox[backward] + plsRcox",
            dataset="Dataset1")


}
