% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Dev.Pred.Category.Sig.R
\name{ML.Dev.Pred.Category.Sig}
\alias{ML.Dev.Pred.Category.Sig}
\title{Title}
\usage{
ML.Dev.Pred.Category.Sig(
  train_data,
  list_train_vali_Data,
  candidate_genes = NULL,
  methods = NULL,
  seed = 5201314,
  cores_for_parallel = 12
)
}
\arguments{
\item{cores_for_parallel}{}
}
\description{
Title
}
