% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_vis.R
\name{roc_vis}
\alias{roc_vis}
\title{Title}
\usage{
roc_vis(
  object,
  model_name,
  dataset,
  order = NULL,
  dataset_col = NULL,
  anno_position = NULL,
  year
)
}
\arguments{
\item{year}{}
}
\description{
Title
}
