% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_sur.R
\name{rs_sur}
\alias{rs_sur}
\title{Title}
\usage{
rs_sur(
  object,
  model_name,
  dataset,
  cutoff = NULL,
  conf.int = F,
  median.line = NULL,
  color = NULL,
  xlab = NULL,
  pval.coord = NULL
)
}
\arguments{
\item{pval.coord}{}
}
\description{
Title
}
