% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_score.R
\name{signature_score}
\alias{signature_score}
\title{Title}
\usage{
signature_score(
  res.by.ML.Dev.Prog.Sig,
  inputmatrix.list,
  signature = IOBR::signature_collection,
  estima_method = "zscore",
  diff.test = T,
  correlation.test = T,
  cor.method = "pearson",
  cor.siganatue = c("CD_8_T_effector", "DDR")
)
}
\arguments{
\item{cor.siganatue}{}
}
\description{
Title
}
