% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TME_deconvolution_all.R
\name{TME_deconvolution_all}
\alias{TME_deconvolution_all}
\title{Title}
\usage{
TME_deconvolution_all(
  inputmatrix.list,
  deconvolution_methods = "ALL",
  microarray_names = NULL
)
}
\arguments{
\item{microarray_names}{}
}
\description{
Title
}
