% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_dis_all.R
\name{auc_dis_all}
\alias{auc_dis_all}
\title{Title}
\usage{
auc_dis_all(
  object,
  color = NULL,
  dataset_col = NULL,
  dataset,
  validate_set,
  order = NULL,
  year,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
