% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_dis_select.R
\name{auc_dis_select}
\alias{auc_dis_select}
\title{Title}
\usage{
auc_dis_select(
  object,
  model_name,
  dataset_col = NULL,
  dataset,
  order = NULL,
  year
)
}
\arguments{
\item{year}{}
}
\description{
Title
}
