% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_auc_pre.prog.sig.R
\name{cal_auc_pre.prog.sig}
\alias{cal_auc_pre.prog.sig}
\title{Title}
\usage{
cal_auc_pre.prog.sig(
  use_your_own_collected_sig,
  collected_sig_table,
  type.sig,
  list_input_data,
  AUC_time = 1,
  auc_cal_method = "KM"
)
}
\arguments{
\item{auc_cal_method}{}
}
\description{
Title
}
