% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_auc_previous_immunotherapy_sig.R
\name{cal_auc_previous_sig}
\alias{cal_auc_previous_sig}
\title{Title}
\usage{
cal_auc_previous_sig(
  list_train_vali_Data,
  seed = 5201314,
  train_data,
  cores_for_parallel = 12
)
}
\arguments{
\item{cores_for_parallel}{}
}
\description{
Title
}
