% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{cindex_dis_all}
\alias{cindex_dis_all}
\title{Title}
\usage{
cindex_dis_all(
  object,
  color = NULL,
  dataset_col = NULL,
  validate_set,
  order = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
