% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_feature_sur.R
\name{core_feature_sur}
\alias{core_feature_sur}
\title{Title}
\usage{
core_feature_sur(
  gene,
  InputMatrix,
  dataset,
  cutoff = NULL,
  conf.int = F,
  median.line = NULL,
  color = NULL,
  xlab = NULL,
  pval.coord = NULL
)
}
\arguments{
\item{pval.coord}{}
}
\description{
Title
}
