% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Corefeature.Prog.Screen.R
\name{ML.Corefeature.Prog.Screen}
\alias{ML.Corefeature.Prog.Screen}
\title{Title}
\usage{
ML.Corefeature.Prog.Screen(
  InputMatrix,
  candidate_genes,
  mode = NULL,
  seed = NULL,
  single_ml = NULL,
  nodesize = 5
)
}
\arguments{
\item{nodesize}{}
}
\description{
Title
}
