% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Dev.Prog.Sig.R
\name{ML.Dev.Prog.Sig}
\alias{ML.Dev.Prog.Sig}
\title{Title}
\usage{
ML.Dev.Prog.Sig(
  train_data,
  list_train_vali_Data,
  candidate_genes = NULL,
  mode = NULL,
  single_ml = NULL,
  alpha_for_Enet = NULL,
  direction_for_stepcox = NULL,
  double_ml1 = NULL,
  double_ml2 = NULL,
  nodesize = NULL,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
