% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_vis_category.R
\name{roc_vis_category}
\alias{roc_vis_category}
\title{Title}
\usage{
roc_vis_category(
  object,
  model_name,
  dataset,
  order = NULL,
  dataset_col = NULL,
  anno_position = NULL
)
}
\arguments{
\item{anno_position}{}
}
\description{
Title
}
